amd64: Change "pointer" fields in java objects from int to long. (Bugfix)

+++ jbinding-cpp/JavaToCPP/JavaToCPPInArchiveImpl.cpp
@@ -89,7 +89,7 @@ static CPPToJavaInStream * GetInStream(JNIEnv * env, jobject thiz)
 
 	localinit(env, thiz);
 
-	pointer = env->GetIntField(thiz, g_InStreamAttributeFieldID);
+	pointer = env->GetLongField(thiz, g_InStreamAttributeFieldID);
 
 	if (!pointer)
 	{
